+++ OSMSurveyorFrameworkTests/Factories/Node+MakeNode.swift
@@ -10,9 +10,9 @@ import Foundation
 @testable import OSMSurveyorFramework
 
 extension Node {
-    static func makeNode(id: Int64 = 1,
+    static func makeNode(id: Int = 1,
                          coordinate: Coordinate = Coordinate(latitude: 53.5, longitude: 9.2),
-                         version: Int64 = 2,
+                         version: Int = 2,
                          tags: [String: String] = ["amenity": "bench"]) -> Node {
         return Node(id: id,
                     coordinate: coordinate,
